backup_before_reorganization_20250723_013430/AOI_Base_Files_Schwabot/backup_before_schwabot_20250717_232349/requirements.txt
# Schwabot Trading System Requirements
# Core dependencies for production-ready crypto trading bot

# Core Python packages
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Trading and financial analysis
ccxt>=2.0.0
yfinance>=0.1.70
ta-lib>=0.4.24
backtrader>=1.9.76.123

# Machine learning and AI
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.15.0

# Quantum computing (optional)
pennylane>=0.20.0

# Web framework and API
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.26.0
aiohttp>=3.8.0

# Database and caching
redis>=4.0.0
sqlalchemy>=1.4.0
sqlite3

# Configuration and environment
pyyaml>=6.0
python-dotenv>=0.19.0
configparser

# Logging and monitoring
loguru>=0.6.0
psutil>=5.8.0

# Security and encryption
cryptography>=3.4.0
hashlib

# Data processing and visualization
plotly>=5.0.0
dash>=2.0.0
bokeh>=2.4.0

# Time series analysis
statsmodels>=0.13.0
arch>=5.0.0

# GPU acceleration (optional)
cupy-cuda11x>=10.0.0; sys_platform == "linux"
cupy-cuda12x>=10.0.0; sys_platform == "linux"

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.16.0
flake8>=4.0.0
black>=21.0.0

# System utilities
psutil>=5.8.0
GPUtil>=1.4.0

# Network and communication
websockets>=10.0
asyncio-mqtt>=0.11.0

# File handling
pathlib
shutil

# Mathematical libraries
sympy>=1.9.0
mpmath>=1.2.0

# Signal processing
scipy.signal
scipy.fft

# Optimization
scipy.optimize
cvxpy>=1.2.0

# Time handling
python-dateutil>=2.8.0
pytz>=2021.3

# JSON and data serialization
ujson>=4.0.0
msgpack>=1.0.0

# Async support
asyncio
aiofiles>=0.8.0

# Memory management
psutil>=5.8.0
memory-profiler>=0.60.0

# Performance monitoring
py-spy>=0.3.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0 